% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{statisticalAnalysis}
\alias{statisticalAnalysis}
\title{statisticalAnalysis
Provides automated statistical analysis of a CSV file}
\usage{
statisticalAnalysis()
}
\value{
pdf with analysis results
}
\description{
statisticalAnalysis
Provides automated statistical analysis of a CSV file
}
\examples{
statisticalAnalysis()
}
